--- a/optuna/samplers/_tpe/sampler.py
+++ b/optuna/samplers/_tpe/sampler.py
@@ -376,13 +376,16 @@ class TPESampler(BaseSampler):
 
         # We divide data into below and above.
         indices_below, indices_above = _split_observation_pairs(scores, self._gamma(n))
-        below = _build_observation_dict(values, indices_below)
-        above = _build_observation_dict(values, indices_above)
+        # `None` items are intentionally converted to `nan` and then filtered out.
+        # For `nan` conversion, the dtype must be float.
+        config_values = {k: np.asarray(v, dtype=float) for k, v in values.items()}
+        below = _build_observation_dict(config_values, indices_below)
+        above = _build_observation_dict(config_values, indices_above)
 
         # We then sample by maximizing log likelihood ratio.
         if study._is_multi_objective():
             weights_below = _calculate_weights_below_for_multi_objective(
-                values, scores, indices_below
+                config_values, scores, indices_below
             )
             mpe_below = _ParzenEstimator(
                 below, search_space, self._parzen_estimator_parameters, weights_below
@@ -420,12 +423,15 @@ class TPESampler(BaseSampler):
             )
 
         indices_below, indices_above = _split_observation_pairs(scores, self._gamma(n))
-        below = _build_observation_dict(values, indices_below)
-        above = _build_observation_dict(values, indices_above)
+        # `None` items are intentionally converted to `nan` and then filtered out.
+        # For `nan` conversion, the dtype must be float.
+        config_values = {k: np.asarray(v, dtype=float) for k, v in values.items()}
+        below = _build_observation_dict(config_values, indices_below)
+        above = _build_observation_dict(config_values, indices_above)
 
         if study._is_multi_objective():
             weights_below = _calculate_weights_below_for_multi_objective(
-                values, scores, indices_below
+                config_values, scores, indices_below
             )
             mpe_below = _ParzenEstimator(
                 below,
@@ -700,12 +706,8 @@ def _split_observation_pairs(
 
 
 def _build_observation_dict(
-    config_vals: Dict[str, List[Optional[float]]], indices: np.ndarray
+    config_values: Dict[str, np.ndarray], indices: np.ndarray
 ) -> Dict[str, np.ndarray]:
-
-    # `None` items are intentionally converted to `nan` and then filtered out.
-    # For `nan` conversion, the dtype must be float.
-    config_values = {k: np.asarray(v, dtype=float) for k, v in config_vals.items()}
 
     observation_dict = {}
     for param_name, param_val in config_values.items():
@@ -762,7 +764,7 @@ def _solve_hssp(
 
 
 def _calculate_weights_below_for_multi_objective(
-    config_vals: Dict[str, List[Optional[float]]],
+    config_vals: Dict[str, np.ndarray],
     loss_vals: List[Tuple[float, List[float]]],
     indices: np.ndarray,
 ) -> np.ndarray:
@@ -772,7 +774,7 @@ def _calculate_weights_below_for_multi_objective(
     # misses the one trial, then the other parameter must miss the trial, in this call of
     # `sample_relative`.
     # In the call of `sample_independent`, we only have one parameter so the logic makes sense.
-    cvals = np.asarray(list(config_vals.values())[0])[indices]
+    cvals = list(config_vals.values())[0][indices]
 
     # Multi-objective TPE does not support pruning, so it ignores the ``step``.
     lvals = np.asarray([v for _, v in loss_vals])[indices]