@not522 2021-07-21
The improvement mainly comes from the change of `parzen_estimator.py`. The following is the benchmark result for only that change.


<details>
<summary>suggest_categorical</summary>

| #params | #trials | time(sec) | time/trial(ms) | time/trial/param(ms) |
| ------- | ------- | --------- | -------------- | -------------------- |
| 1 | 1 | 0.00 | 0.39 | 0.39 |
| 1 | 10 | 0.00 | 0.14 | 0.14 |
| 1 | 100 | 0.05 | 0.46 | 0.46 |
| 1 | 1000 | 1.89 | 1.89 | 1.89 |
| 2 | 1 | 0.00 | 0.33 | 0.17 |
| 2 | 10 | 0.00 | 0.28 | 0.14 |
| 2 | 100 | 0.09 | 0.86 | 0.43 |
| 2 | 1000 | 3.48 | 3.48 | 1.74 |
| 4 | 1 | 0.00 | 0.35 | 0.09 |
| 4 | 10 | 0.00 | 0.32 | 0.08 |
| 4 | 100 | 0.16 | 1.55 | 0.39 |
| 4 | 1000 | 6.91 | 6.91 | 1.73 |
| 8 | 1 | 0.00 | 0.92 | 0.12 |
| 8 | 10 | 0.01 | 0.62 | 0.08 |
| 8 | 100 | 0.31 | 3.13 | 0.39 |
| 8 | 1000 | 13.80 | 13.80 | 1.72 |
| 16 | 1 | 0.00 | 1.02 | 0.06 |
| 16 | 10 | 0.01 | 1.29 | 0.08 |
| 16 | 100 | 0.57 | 5.72 | 0.36 |
| 16 | 1000 | 28.66 | 28.66 | 1.79 |
| 32 | 1 | 0.00 | 1.78 | 0.06 |
| 32 | 10 | 0.02 | 2.01 | 0.06 |
| 32 | 100 | 1.16 | 11.64 | 0.36 |
| 32 | 1000 | 58.43 | 58.43 | 1.83 |

</details>